Step1: Move 1D solver and agent GUI modules into pilot subpackages
The pilot package keeps every module in one flat directory, so the 1D
CESE solver demos and the agent console sit beside the application
shell, the canvas, and the dock panels with nothing but file names to
group them.  Give the two least-connected families their own homes as
the first step of the pilot package reorganization: the Euler,
Burgers, and linear-wave demos move to `solvcon/pilot/onedim/`, and
the agent console moves to `solvcon/pilot/agent/`.

Each new subpackage gains an `__init__.py` that guards the
Qt-dependent modules behind the `_pilot_core` enable flag and
re-exports the public app classes, following the model of the existing
`_euler` and `airfoil` subpackages.  The moved modules only see their
relative imports gain a level; the public `solvcon.pilot` surface does
not change.
