+++ setup.py
@@ -74,7 +74,9 @@ def main():
             'solvcon.onedim',
             'solvcon.pilot',
             'solvcon.pilot._euler',
+            'solvcon.pilot.agent',
             'solvcon.pilot.airfoil',
+            'solvcon.pilot.onedim',
             'solvcon.plot',
             'solvcon.profiling',
             'solvcon.track',
