+++ solvcon/pilot/agent/_agent_gui.py
@@ -18,8 +18,8 @@ from PySide6.QtWidgets import (QWidget, QVBoxLayout, QHBoxLayout, QDockWidget,
                                QLabel, QComboBox, QTextEdit, QLineEdit,
                                QPushButton)
 
-from ..agent import AgentSession, available_backends
-from . import _gui_common
+from ...agent import AgentSession, available_backends
+from .. import _gui_common
 
 __all__ = [  # noqa: F822
     'AgentBackendWorker',
