+++ solvcon/pilot/onedim/_linear_wave.py
@@ -4,8 +4,8 @@
 
 import numpy as np
 
-from . import _base_app
-from ..onedim import linear_scalar
+from .. import _base_app
+from ...onedim import linear_scalar
 
 
 class LinearWave:
